+++ include/sub2ind.h
@@ -1,6 +1,8 @@
 #ifndef SUB2IND_H
 #define SUB2IND_H
 
-int sub2ind(int i, int j, int num_cols);
+#include <stdlib.h>
+
+size_t sub2ind(int i, int j, int num_cols);
 
 #endif
