+++ src/mesh_to_vector.c
@@ -1,7 +1,9 @@
+#include <stdlib.h>
+
 #include "mesh.h"
 #include "vector.h"
 
-void mesh_to_vector(const Mesh *mesh, int index, Vector *v) {
+void mesh_to_vector(const Mesh *mesh, size_t index, Vector *v) {
     v->x = mesh->x[index];
     v->y = mesh->y[index];
     v->z = mesh->z[index];
