+++ src/sub2ind.c
@@ -1,3 +1,5 @@
-int sub2ind(int i, int j, int num_cols) {
-    return i * num_cols + j;
+#include <stdlib.h>
+
+size_t sub2ind(int i, int j, int num_cols) {
+    return ((size_t) i) * num_cols + j;
 }
